Use Rcpp export rather than .Call
This is a maintenance issue. The name generated by Rcpp for the C call isn't guaranteed to be stable across versions. Also, other options set, such as RNG state, in Rcpp attributes, can result in the RcppExport function having a few protective/state-setting components.  So it's best to use it rather than skip it.